Rename thorchainDa.ts to swapkit.ts
